Add a runtests.py for easily running the test suite.
This adds a tests/runtests.py script for easily running the test suite.
All that's needed is a standard checkout. Right now, this only tests with
a sqlite database, but will be updated soon to allow testing with multiple
database types.

This also includes a tests/run-pyflakes.py for easily running pyflakes on
the codebase and ignoring certain types of warnings.