Normalize ManyToManyField 'through' tables for the test suite.
The test suite didn't know how to handle the new required ManyToManyField
'through' tables and so it never normalized the names like it did the
parent models, or added it to its own model cache for the test signature
functions.

We now apply the same basic logic as with the parent models, taking into
account that the old parent model names formed the basis of the 'through' model
names, and replacing those parts with the new names.