Add support for Django 1.2.
This introduces support for Django 1.2, while maintaining compatibility
with Django 1.1.

Much of the internal database functionality changes in Django 1.2 that
Django Evolution relied on. The primary breakage is that in 1.2, the
ManyToManyField 'through' tables were no longer created behind the scenes.
It used to be that unless a through= parameter was passed, it would be set
to None and the table would be created behind the scenes. Now, it's always
set, and that table is taken into account when generating the SQL and
signatures we see.

They would, however, be created with the wrong name in our test suite, given
our crazy TestModel renaming logic. We now handle creating these tables
properly, referencing the correcet fields, and also iterate through them
and rename the right keys for our test infrastructure.

These also affected ManyToManyField addition code, table destruction,
db_table property change, primary key renaming, and more, due to constraints
being added in the PostgreSQL database backend. We now iterate through when
necessary and remove these constraints before making the change, and then
re-adding them with the new names.

Django 1.2 also caches its list of app models, which it didn't used to do.
This broke our test infrastructure, as we assumed we could always just modify
this list. We also assumed it could be in whatever order we wanted. This isn't
true anymore, though. We now take care to clear the cache when registering
models for the test infrastructure and to put them in the right order.

We do a better job of preserving order in many places now. Given the
constraint requirements with the 'through' tables, we needed to have
constraints added and deleted in the proper order, which meant having
models added/deleted in the proper order. For this, we're using Django's
SortedDict in more places now.

This has been tested with Django 1.1 and 1.2 on Python 2.5 and 2.6 with
SQLite, MySQL, and PostgreSQL.