Fix incorrect defaults on SQLite when adding null fields.
On SQLite, adding a null field without a default value would cause the
field name to be the default. This was due to attempting to select the
field name from the temporary table, but since the table didn't exist,
the field name itself was being used as the value.

We are now more explicit about the fields being selected and populated.
We have two lists, and no longer assume both are identical. We also
use NULL columns for temporary table fields unconditionally.

Thanks to smileychris for working with me on the fix for this long ago.

Fixes bug #49