Fix --hint output, and check for multi-DB support in more places.
The --hint output was mistakenly warning that the evolutions it just
provided could not apply. This was due to a code block being accidentally
indented one too many levels. It's now fixed and the hint output is
correct.

The evolve command is now more careful to check for multi-db support in
more places.