Reduce redundant function calls when possible.
This change reduces the number of redundant function calls for database
operations. The new multi-DB patch updated many call sites to call database
functions using the 'using=' argument if multi DB support was enabled,
or without it otherwise. This led to a lot of excess calls.

Now in cases where we need to perform the call with different parameters,
we now create a using_args dictionary, which will be populated with a
proper using= if multi-DB support is provided. We then use that for each
call. For Django 1.1, this will result in no extra arguments, but for
Django 1.2, this will result in a using= parameter.