Add support for automatically evolving changes to built-in Django models.
Django 1.3 alpha changed the database schema for Session, setting db_index
to True on the expire_date field. This requires a database schema, but there
was nothing in place for actually evolving this schema.

We now have a place for providing evolution data for built-in Django models.
These now live in django_evolution/builtin_evolutions. If we find a schema
change needed that is provided in this list, we include it. The only one
in there so far is the db_index change, which we use introspection to
determine whether or not to provide.