Fix multi-database support and the unit tests for checking it.
Multi-database support wasn't fully-featured, as the evolver wasn't
being passed the correct database. This was verified with unit tests
when two databases were of different types (the generated SQL would be
wrong).

In practice, if the databases were of the same type, multi-database
support would work, but not if they were of different types.

Now the unit tests pass when mixing databases.