Fix adding a non-null field with a default value on PostgreSQL.
Adding a non-null field with a default value on PostgreSQL would result in
a failure if the table had any existing data. This was due to an ALTER TABLE
followed by an UPDATE in the same transaction, which will complain about
pending trigger events.

Instead of a ALTER TABLE + UPDATE, we now do two ALTER TABLEs. One creates
the field, with a default value, and the other drops the default value.
This is an alternative to populating the existing tables using UPDATE.
We end up dropping it in order to stay consistent with Django, which
doesn't add default value information to fields at the database level.

Fixes bugs #58 and #74.