Install a baseline evolution history for new apps.
When upgrading an older database using Django Evolution when a new model
has been added and subsequent evolutions were made on that model, the upgrade
would fail. It would attempt to apply those evolutions on that model, which,
being newly created, would already have those new field changes.

Now, like with an initial database, we install a baseline evolution history
for any new apps. This will ensure that those evolutions aren't applied to
the models in that app.