Fix compatibility with Django SVN in the unit tests.
In Django SVN r16053, get_model() and get_models() only return installed
modules by default. This is calculated in part by a new AppCache.app_labels
dictionary, along with an existing AppCache.app_store, neither of which
we properly populated.

We now set both of these (though, app_labels only on versions of Django
that have it). This allows the unit tests to pass, both with older versions of
Django and Django SVN.