Associate the correct version when installing app baselines.
The new code for installing a baseline for new apps in 0.6.4 was associating
the wrong Version model with the Evolution. This doesn't appear to cause any
real-world problems, but it does make it harder to debug things when evolutions
go wrong.

Instead of creating the baselines as we find new apps, we instead put them
in a list and then create them once the new Version entry is created, in order
to associate it with that Version.