Generate more accurate sample evolutions with non-Django DB fields.
If a project introduces a new field that isn't supplied by Django's
core set of database fields, a non-functioning evolution would be
generated. The assumption was that every type of field was provided in
django.db.models.

Now the generated evolution includes proper imports for such fields. It
also groups them according to PEP-8 standards to the best of its
ability.