Don't attempt to install evolutions for apps without models.
Installing a baseline evolution for apps without models would fail. Now
we tell Django that we'll accept a None return value for such apps, and
just skip over them.

This prevents issues when an app is in INSTALLED_APPS for things such as
templatetags, when they don't have any models defined.