Add two new management commands, list-evolutions and wipe-evolution.
list-evolutions lists all applied evolutions. It can take one or more
app labels, and will restrict the output to those apps.

wipe-evolution will wipe one or more evolutions from the database. This
should only be used if absolutely necessary, and can cause problems. It
is useful if there's some previously applied evolutions getting in the
way, which can happen if a person is uncareful with downgrading and
upgrading again.