Update the README to remove some warnings and point to South.
Most projects would be best served by starting with South, instead of
Django Evolution, since the momentum is going there. That doesn't mean
development is stopped here, though.

Removed the warning about not being used on production servers, because
in reality, it's used on thousands of them, largely without problems.