Filter unchanged models when determining mutations to apply.
Django Evolution has had a problem concerning mutations to models that
were just applied as a baseline.

Take, for example, the case where someone updates from App v1.0 to 3.0.
In 2.0, which was never installed, a model was added, and in 3.0 a
mutation was added.

Previously, Django Evolution would attempt to apply a baseline for 3.0
(installing a signature as it is in 3.0), and then attempt to apply
mutations found in 3.0.

The mutations would fail, since the fields were already there. Even
worse, older mutations may mess things up.

Now we filter the list of mutations and only apply those to model
signatures that have changed since the last time a signature was
generated. If a baseline was just installed, then any new mutations
should be skipped. If it's an existing signature, though, any new
mutations will be applied fine.