Add easy testing against different types of databases.
runtests.py now takes a parameter for a database entry it can run unit
tests against. By default, there is one: "sqlite3".

Additional entries can be added by creating a test_db_settings.py in the
tests/ directory. There's a template for a sample configuration for
MySQL (with MyISAM and INNODB) and PostgreSQL.