Require ChangeMetas for existing models containing unique_together.
Django Evolution previously stored but otherwise ignored unique_together
entries, meaning that older values could not be trusted to have been
applied to the database It could be wildly inconsistent with what's
actually there.

This change considers an older unique_together to be invalidated,
requiring a ChangeMeta mutation, if from Django Evolution 0.6.x or
older.

There's a new flag stored in the signature for each table called
__unique_together_applied. This is used to indicate whether the
field has been applied to the database (or, in the case of an empty
unique_together, at least considered).

When diffing, two identical unique_together values are considered
to be unequal if the older signature doesn't have this new flag.

Projects with existing unique_together values will need to provide
a ChangeMeta for each table. Samples are shown when using evolve --hint.

This also provides some built-in evolutions for the main affected Django
models.