Fix return types with no-op column renames on MySQL.
When attempting to generate the SQL for renaming columns on MySQL, we'd
return None in the cases where the old and new column names were the
same, or when attempting to rename a ManyToManyField.

None is not a valid return type, and results in a crash later when
iterating through the resulting SQL strings. Instead, we just need to
return an empty list.