Fix bad return types in Postgres and common DB code.
There were a few other call sites that returned None instead of an empty
list as resulting SQL. None is no longer valid, and was unexpected,
which caused issues when building the evolutions to apply.

All call sites that were returning None that should return a list now do
so.