Prevent attempting to create/delete indexes unnecessarily.
If the db_index parameter was set to True and an index already existed
in the databse, or it was set to False and an index didn't exist in the
database, we'd still generate the SQL for the index change, leading to
errors.

We now do a better job of detecting whether we actually need to
create/delete the index. Unit tests were added to ensure this works
across databases.

Testing Done:
Unit tests pass on all databases.

Solved a problem I was hitting with a real-world database.

Reviewed at https://reviews.reviewboard.org/r/7164/