Release Django Evolution 0.6.10.
If you're reading this, you may be wondering what's in this release,
given the commit history. Well, just a README update, actually. Surely
that's not worthy of a release, you may think. Well, it's not, but
here's the thing...

The 0.6.9 build was once uploaded to PyPI, but we removed it in favor of
using an external index, which is no longer supported by PyPI. So we
wanted to move back, but PyPI does not allow existing filenames to be
released. So, 0.6.9 is basically gone forever...

So in order to make a 0.6.x release installable via easy_install or pip,
we must produce a new release.

Thanks, PyPI.