Fix a regression with renaming models.
A recent change (made since the last public release) implemented fixes
for renaming models. However, it resulted in a breakage in the case
where a baseline contained the signature for the renamed model along
with the original model. In this case, since the new name was found in
the signature, the mutation was considered unnecessary and removed, but
this wasn't actually correct, leading to an evolution failure.

We now only remove that mutation if the new name exists and the old one
does not, fixing renames for models.

Testing Done:
Unit tests pass.

Tested against a database schema we had that failed to upgrade after
the recent fix for renames.

Reviewed at https://reviews.reviewboard.org/r/8310/