Improve the information shown for the evolution process.
We've seen some database corruption issues in the past due to people
thinking the evolution process has stalled and hitting Control-C. In an
attempt to prevent this, a few additions are being made to the evolution
output.

First, in the confirmation message, we tell users to back up their
database. Often times, admins or new developers don't think to do this,
and break something. This message is only shown when interactive mode is
on.

We now specifically say it may take some time and not to cancel the
upgrade, in an attempt to get people to not hit Control-C.

As we execute SQL statements for an app, we list the app currently being
evolved. This gives a sense of progress, and is useful when diagnosing
where something went wrong.

When something *does* go wrong, we also now have some more data to work
with. We list the app being evolved, the SQL statement that failed, and
the database error. Through this, it's easier to track down what went
wrong and where.

Testing Done:
Ran some evolutions, successful and with (simulated) failures. Saw the
progress information and the extra failure information.

Reviewed at https://reviews.reviewboard.org/r/8311/