Add utility functions for getting an app's name and label.
The django-evolution codebase is littered with snippets of code for
calculating the label and name for an app. This is kind of annoying and
not very future-proof. This change replaces all those with calls to new
utility functions designed to return that information.

Testing Done:
Tested listing evolutions, generating evolution hints, and applying
evolutions. All continued to work.

Reviewed at https://reviews.reviewboard.org/r/8633/