Inspect the schema and install baselines after running manage.py migrate.
Newer versions of Django don't have syncdb, and therefore don't have the
post_syncdb signal we use in order to install model baselines and to
determine if we need to recommend installing evolutions.

This change updates our code to listen for a completed migration of an
app when on a newer version of Django. The common bits of code have been
moved into a common method shared by the signal handles for both syncdb
and migrate commands. This also cleans up and documents some of the code
in this path.

Testing Done:
Been running this on my local development trees for a while, and tested
it with the Django 1.7+ unit tests in consuming projects.

Reviewed at https://reviews.reviewboard.org/r/8804/