Add support for Django 1.11.
This introduces support for all database types on Django 1.11. The major
differences in this version have to do with how index names are
generated (again), how Postgres constraint-related SQL is generated
(again), and the introduction of yet another way of setting indexes on a
model.

Django 1.11 allows explict indexes to be defined by setting a model's
`Meta.indexes` to a list of `models.Index` instances, which can span
fields. This change introduces support for representing this in the
signature, to set new indexes in `ChangeMeta`, and to evolve these
changes.

It also handles the other compatibility issues for index names (which,
unfortunately, are not consistent between the new `models.Index`
instances and the older indexes) and changes to generated SQL.

Testing Done:
Unit tests pass for all supported database types on Django 1.6 through
Django 1.11.

Reviewed at https://reviews.reviewboard.org/r/9560/