Exclude database test data files from the test runner.
The test runner was attempting to discover tests in the
`django_evolution.tests.db.*` package, which triggered database
connections and caused errors with testing if a particular database
wasn't configured or running.

We now use the `nose-exclude` plugin and a configuration option to tell
nose not to look for any tests there. This plugin is required due to
the fact that `--exclude` and `-ignore-files` won't work on directories.

Testing Done:
Added some asserts to the test data files to simulate the behavior
being seen in CI. Saw the failures prior to the new option and plugin.
After, those modules weren't being imported and all tests passed.

Reviewed at https://reviews.reviewboard.org/r/9838/